types: Exclude Cabal targets when `stack.yaml` exists
Haskell Stack looks for a file `stack.yaml` to find its configuration.
When using `projection-multi-compile` in a project that has both a
`stack.yaml` file and a (usually generated) `.cabal` file,
`projection-multi-compile` will currently display both Cabal and Stack
targets.  It looks like this is not the intention, as Cabal targets
are excluded when a `stack.yml` file exists.

This patch makes Cabal targets be excluded when a `stack.yaml` file
exists as well.  A cursory examination of the source code to Stack and
its documentation makes it seem that Stack will only look for
`stack.yaml` files now.  There seem, however, to be plenty of projects
that still use `stack.yml` files, perhaps with an older version of
Stack, so this patch will not break them.