types/haskell-stack: Add `compile-multi` support for Stack
Stack provides a `stack ide` command that outputs either a list of all
the packages in a project, or all the targets you can choose to build in
a project.  This patch extends the `haskell-stack` build type to call
this command, parse its output, and show each package and target as a
`compile-multi` target, with the ability to build, test, benchmark,
clean, and run packages, and more.  This patch additionally adds other
globally available targets to benchmark, generate documentation, and
clean the build directory.